feat: button 중복 제거
