+++ src/css/page/FacilitiesMain.module.css
@@ -1,13 +1,18 @@
 .facilitiesMain,
 .main {
+  width: 1160px;
+  position: relative;
   display: flex;
   flex-direction: column;
   align-items: flex-start;
   justify-content: flex-start;
-}
-.main {
   align-self: stretch;
   max-width: 100%;
+  width: 1160px;
+  margin: auto;
+  background-color: var(--default-white);
+  overflow: hidden;
+  gap: 254px;
 }
 .facilitiesMain {
   width: 100%;
@@ -17,6 +22,55 @@
   gap: 254px;
   line-height: normal;
   letter-spacing: normal;
+}
+.team1Wrapper {
+  align-self: stretch;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  justify-content: flex-end;
+  padding: 0 var(--padding-10xs) 0 0;
+  box-sizing: border-box;
+  max-width: 100%;
+}
+.hero{
+  width: 1160px;
+  margin: auto;
+}
+
+.content,
+.meetOurStaff {
+  display: flex;
+  justify-content: flex-start;
+  max-width: 100%;
+}
+.meetOurStaff {
+  flex: 1;
+  flex-direction: column;
+  align-items: flex-end;
+  gap: var(--gap-20xl);
+}
+.content,
+.hero {
+  align-items: flex-start;
+}
+.content {
+  width: 616px;
+  flex-direction: row;
+  padding: 0 var(--padding-31xl);
+  box-sizing: border-box;
+  font-size: var(--font-size-3xl);
+  color: var(--medium);
+  font-family: var(--font-poppins);
+}
+.hero {
+  align-self: stretch;
+  flex-direction: column;
+  gap: var(--gap-12xl);
+  text-align: left;
+  font-size: var(--font-size-34xl);
+  color: var(--default-white);
+  font-family: var(--font-roboto);
 }
 @media screen and (max-width: 725px) {
   .facilitiesMain {
