+++ src/pages/FacilitiesMain.jsx
@@ -1,12 +1,17 @@
-import Hero from "../components/Hero";
+import Team1 from "../components1/Team1";
 import styles from "../css/page/FacilitiesMain.module.css";
 
 const FacilitiesMain = () => {
+  
   return (
-    <div className={styles.facilitiesMain}>
-      <section className={styles.main}>
-        <Hero />
-      </section>
+    <div className={styles.main}>
+        <div className={styles.content}>
+          <div className={styles.meetOurStaff}>
+            <Team1 />
+            <Team1 />
+          </div>
+        </div>
+      
     </div>
   );
 };
