+++ src/pages/LandingPage.jsx
@@ -1,7 +1,7 @@
 import Content from "../components1/Content";
 import EventCalendar from "../components1/EventCalendar";
-import Notice from "../components/Notice";
-import Gallery from "../components/Gallery";
+import Notice from "../components1/Notice";
+import Gallery from "../components1/Gallery";
 import styles from "../css/page/LandingPage.module.css";
 
 const LandingPage = () => {
