+++ src/pages/Notice1.jsx
@@ -1,14 +1,7 @@
-import { useCallback } from "react";
-import { useNavigate } from "react-router-dom";
-import ContentLeft from "../components/ContentLeft";
+import ContentLeft from "../components1/ContentLeft";
 import styles from "../css/page/Notice1.module.css";
 
 const Notice1 = () => {
-  const navigate = useNavigate();
-
-  const onImageClick = useCallback(() => {
-    navigate("/");
-  }, [navigate]);
 
   return (
     <div className={styles.notice}>
